Move logic into utils file, add EXTREMELY basic RTL test
